Contracts: make EIP 712 BT constants private, consistent with examples and to keep interface lean.
